Share head-identity resolution with inlay hints; drive element retrieval from the registry
`head_identity` now owns the `namespace import` scan too, so one function
answers the whole identity question, and `inlay_hints` resolves its head
through it before asking for a format family — an aliased/renamed head
hints like the original and a shadowed one hints not at all.

`collection_head_element_classes` reads
`ReturnElements::ElementOf` through `resolve_call` instead of matching
`dict get` / `lindex` by name, so the qualified spellings resolve too.
